rename libaec CCSDS backend module
